% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_vcdb}
\alias{collapse_vcdb}
\title{Collapse the VCDB data frame into a more conventional "tidy" data frame}
\usage{
collapse_vcdb(vcdb)
}
\value{
a collapsed data frame more suited for tidyverse-esque EDA tasks
}
\description{
Shrink the dimension of a VCDB data frame by using a representative value for each enumeration.
This function results in some loss of fidelity, a reasonable trade-off for the convenience we get
from a "tidy" data frame.
}
\details{
The function handles logical enumerations (TRUE/FALSE) differently from factor enumerations as
well as numeric enumerations. The resulting data frame (output) contains new variables not in
the original VCDB that stores the "representative" value for each incident across each
enumeration group
}
\examples{
collapse_vcdb(vcdb)
}
