% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{determine_impact}
\alias{determine_impact}
\title{Determine the value for impact.overall_amount}
\usage{
determine_impact(data)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format.}
}
\value{
a data frame
}
\description{
For each row, find a value that represent the impact.
}
\examples{
determine_primary(vcdb)
}
