% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{determine_primary}
\alias{determine_primary}
\title{Determine the primary value from each enumeration group (numeric)}
\usage{
determine_primary(data, string)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format}
}
\value{
a character vector of the values that represent that enumeration for each numeric observation
}
\description{
For each row, find a value that represent the enumeration. If all values in the group is NA, replace
with "Unknown" variable
}
\examples{
determine_primary(vcdb, "action.hacking.variety")
table(determine_primary(vcdb, c("action.hacking.result", "action.hacking.variety")))
}
