% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{enum2grid}
\alias{enum2grid}
\title{Generate a 2x2 ggplot2 grid}
\usage{
enum2grid(data, params)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format}

\item{params}{Character or vector of characters representing one or more fields of interest (prefix)}
}
\value{
a ggplot object
}
\description{
Create an a4 grid (see Verizon's Data Breach Inves 2012) and return a ggplot object.
Replicates the plota4 function from verisr which referenced deprecated code from data.table.
}
\examples{
enum2grid(vcdb, c("action", "asset.variety"))
enum2grid(vcdb, c("asset.variety", "actor.external.variety"))
}
