% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{extract_enums}
\alias{extract_enums}
\title{Extract all enumerations from a data frame}
\usage{
extract_enums(vcdb)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format.}
}
\value{
a character vector containing the names of all enumerations
}
\description{
Returns a character vector containing the names of all enumerations
}
\examples{
extract_enums(vcdb)
}
