% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{extract_names}
\alias{extract_names}
\title{Extract the name}
\usage{
extract_names(string)
}
\arguments{
\item{string}{Character or vector of characters representing one or more fields of interest (enumeration).}
}
\value{
a character string
}
\description{
Extract the value (name) from the original representation of an enumeration.
}
\examples{
extract_names("action.hacking.result.Exfiltrate")
}
