% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{find_largest}
\alias{find_largest}
\title{Find the largest values}
\usage{
find_largest(data)
}
\arguments{
\item{data}{A \code{data frame} object, where all columns}
}
\value{
a character vector of the values that are largest in its enumeration group
}
\description{
For each row, find a value that represent the enumeration using \code{max.col}. An internal function that
is used in \code{determine_primary} and \code{determine_primaryl}.
}
\examples{
find_largest(vcdb[1:10,57:105])
}
