% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr2.R
\name{getenum_df}
\alias{getenum_df}
\title{Generate a data.frame of counts from one or two enumerations.}
\usage{
getenum_df(data, params)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format}

\item{params}{Character or vector of characters representing one or more fields of interest (enumeration)}
}
\value{
a data frame with count and frequency from enumerating on one or more fields.
}
\description{
Generate counts from the values given one or two enumerations and return both
the count and proportion. Replicates in dplyr the original getenum function
from verisr which referenced deprecated code from data.table.
}
\examples{
getenum_df(vcdb, "asset.variety")
getenum_df(vcdb, c("action", "asset.variety"))
}
