% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr2.R
\name{getenum_df_single}
\alias{getenum_df_single}
\title{Generate a data.frame of counts from an enumeration.}
\usage{
getenum_df_single(data, params)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format.}

\item{params}{Character representing the fieldsof interest (prefix).}
}
\value{
a data frame with count and frequency from enumerating on one field.
}
\description{
Generate counts from the values given one enumeration and return both
the count and proportion. Replicates in base R the getenum.single function
from verisr which referenced deprecated code from data.table.
}
\details{
You almost always want to use \code{getenum_df} instead, which is a more generalized wrapper
that returns an identical data frame when called with one field.
}
\examples{
getenum_df_single(vcdb, "actor.external.variety")
}
\seealso{
\code{\link{getenum_df}} for a more generalized implementation of this function
}
