% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr2.R
\name{getenum_tbl}
\alias{getenum_tbl}
\title{A frequency count table from one or more enumerations}
\usage{
getenum_tbl(data, params)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format}

\item{params}{Character or vector of characters representing one or more fields of interest (enumeration)}
}
\value{
a frequency table from enumerating on the specified fields of interests
}
\description{
Generate counts from the values given one or more enumerations
is immediately preceded by the specified string.
}
\examples{
getenum_tbl(vcdb, c("actor"))
getenum_tbl(vcdb, c("action", "asset.variety"))
}
