% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importveris}
\alias{importveris}
\title{Import all vcdb incidents is a "ready" data frame}
\usage{
importveris(dir, schema = "data/vcdb-merged.json")
}
\arguments{
\item{dir}{The directory to list through. This may be a vector of
directorites, in which case each all the matching files in each
directory will be loaded.}

\item{schema}{a full veris schema with enumerations included}
}
\value{
a data frame
}
\description{
This function is a thin wrapper over the json2veris function. In later versions of vcdb incidents,
the original function may result in a dataframe where one or more of its variables is another level
of nested list object(s). This function eliminates these columns, so they're in a more ready state
for most data analysis tasks.
'
}
\examples{
importveris("~/Datasets/vcdb_small/")
}
