% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr2.R
\name{involving_country}
\alias{involving_country}
\title{Find all incidents involing a specified country, optionally returning only
notes-type columns}
\usage{
involving_country(data, code, notes_only = FALSE)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format.}

\item{code}{Two-letter country codes, e.g "US", "ID" etc}
}
\value{
a data frame containing incidents where the specified country is involved
}
\description{
A function that returns all incidents involving a country, where said country
could be a victim, target, conspirator, owner of assets etc.
}
\details{
When \code{ notes_only } is TRUE, the returned data frame will contain only
columns derived from "notes" relating to the event.
}
\examples{
involving_country(vcdb, "US", notes_only=TRUE)
involving_country(vcdb, "ID")
}
