% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr2.R
\name{notes_only}
\alias{notes_only}
\title{Identify all notes columns}
\usage{
notes_only()
}
\value{
a vector of character containing columns names that are "notes"
}
\description{
Identify all notes columns
}
\examples{
notes_only()

}
\seealso{
\code{\link{involving_country}} where this function is being used
}
