% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{process_log}
\alias{process_log}
\title{Process all logical enumerations in a data frame}
\usage{
process_log(vcdb)
}
\arguments{
\item{data}{A \code{data frame} object, typically converted from the VCDB JSON format.}
}
\value{
a collapsed data frame
}
\description{
Returns a collapsed data frame by selecting a primary value that determine each enumeration
across all logical variables
}
\examples{
process_log(vcdb)
}
