% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verisr2.R
\docType{package}
\name{verisr2}
\alias{verisr2}
\alias{verisr2-package}
\title{verisr2: Convenience functions for exploratory analysis on VERIS database}
\description{
verisr2: Convenience functions for exploratory analysis on VERIS database
}
\section{Motivation}{

The package replicates in base R or dplyr many of the helper functions originally
implemented in the verisr package by Jay Jacobs. The original package by Jay uses
data.table code that is deprecated and no longer works. The author has stated his
desire to one day rewrite these functions in dplyr code but since effort on that
has been stagnant for a few years now (5 years as of writing) this is a simple
attempt to recreate these helper functions in dplyr or base R code.
}

